# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job board API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://express-crud-production-438d.up.railway.app"
    }
  ],
  "paths": {
    "/upload": {
      "post": {
        "summary": "Uploads a file.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "upfile",
            "type": "file",
            "description": "The file to upload."
          }
        ]
      }
    },
    "/misc/parse": {
      "post": {
        "summary": "Allows to upload a file to populate the database (may run out of quota)",
        "tags": ["Miscellany"],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },

        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/misc/common-words": {
      "get": {
        "summary": "Gets the 100 more common words",
        "tags": ["Miscellany"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "post": {
        "summary": "Creates a new category",
        "tags": ["Categories"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Gets all the category names",
        "tags": ["Categories"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "sentences": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sentences": {
      "post": {
        "summary": "Creates a new sentence",
        "tags": ["Sentences"],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },

                  "categories": {
                    "type": "array",
                    "sentences": {
                      "type": "string"
                    },
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "categories": {
                      "type": "array",
                      "sentences": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all sentences",
        "tags": ["Sentences"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "sentences": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "text": {
                        "type": "string"
                      },
                      "categories": {
                        "type": "array",
                        "sentences": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/translate": {
      "post": {
        "summary": "Translate text",
        "description": "Translate input text to the target language using the Google Translate API.",
        "operationId": "translateText",
        "tags": ["Miscellany"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "The text to be translated"
                  },
                  "targetLanguage": {
                    "type": "string",
                    "description": "The language code of the target language (e.g. 'es' for Spanish)"
                  }
                },
                "example": {
                  "text": "Hello world!",
                  "targetLanguage": "es"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "The translated text"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "A description of the error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Translation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "A description of the error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sentences/{id}": {
      "delete": {
        "summary": "Deletes a sentence by id",
        "description": "Deletes a sentence with the specified id from the database.",
        "tags": ["Sentences"],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the sentence to be deleted",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Sentence deleted successfully"
          },
          "404": {
            "description": "Sentence not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Updates a sentence by id",
        "description": "Updates an sentence with the specified id in the database.",
        "tags": ["Sentences"],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sentence updated successfully"
          },
          "404": {
            "description": "Sentence not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The id of the sentence to be updated"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "categories": {
                    "type": "array",
                    "sentences": {
                      "type": "string"
                    },
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}